fix(directive): allow selectors as columnWidth
columnWidth:
"If set to an Element or Selector String, Masonry will use the width of that
element. See Element sizing. Setting columnWidth with element sizing is
recommended if you are using percentage widths."

The previous implementation supported only passing a number that is parsed as
an int before passing to masonry. This prevents using percentage widths and
responsive behaviour.